# public/swagger-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Capstone - API - Documentation",
    "description": "API for user interface development",
    "version": "1.1",
    "contact": {
      "name": "mangox API Support",
      "email": "support@mangox.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/.well-known/oauth-protected-resource": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get___well_known_oauth_protected_resource",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/.well-known/oauth-protected-resource/mcp": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get___well_known_oauth_protected_resource_mcp",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/.well-known/oauth-authorization-server": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get___well_known_oauth_authorization_server",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/oauth/register": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "post__oauth_register",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/oauth/authorize": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__oauth_authorize",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "post__oauth_authorize",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/oauth/token": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "post__oauth_token",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://mangox-api.mangoads.com.vn",
      "description": "Development Server"
    },
    {
      "url": "http://localhost:5555",
      "description": "Development Server"
    }
  ],
...
  "security": [
    {},
    {
      "bearerAuth": []
    }
  ]
}